requirements.txt
dotmap==1.3.30
gymnasium==0.27.1
hydra-core==1.3.2
matplotlib==3.8.2
numpy==1.25.0
omegaconf==2.3.0
torch==2.0.0+cu118
